==> notebooks/09_generate.ipynb <==
import pandas as pd

v4_csv = '../data/cleaned/grocery_stores_cleaned_v4.csv'
v4_json = '../docs/grocery_stores_cleaned_v4.json'  # update path as needed

df = pd.read_csv(v4_csv)
df.to_json(v4_json, orient='records')

print(f"✅ Saved: {v4_json}")
